requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
mss>=9.0.0
pytesseract>=0.3.10
TTS>=0.22.0
rvc-python>=0.1.7
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
keyboard>=0.13.5
torch>=2.0.0
huggingface_hub>=0.22.0
transformers>=4.38.0
sentencepiece>=0.1.99
sacremoses>=0.0.53
pycaw>=20240416
